Create a folder for 'pages' and move Coffee over accordingly
